--- a/src/OpenTelemetry/Metrics/Exemplar/ExemplarMeasurement.cs
+++ b/src/OpenTelemetry/Metrics/Exemplar/ExemplarMeasurement.cs
@@ -51,6 +51,11 @@ internal
     /// <summary>
     /// Gets the measurement tags.
     /// </summary>
+    /// <remarks>
+    /// Note: <see cref="Tags"/> represents the full set of tags supplied at
+    /// measurement regardless of any  filtering configured by a view (<see
+    /// cref="MetricStreamConfiguration.TagKeys"/>).
+    /// </remarks>
     public ReadOnlySpan<KeyValuePair<string, object?>> Tags { get; }
 
     internal int ExplicitBucketHistogramBucketIndex { get; }