… src/OpenTelemetry/.publicApi/Experimental/PublicAPI.Unshipped.txt
… src/OpenTelemetry/CHANGELOG.md
… src/OpenTelemetry/Metrics/Exemplar/Exemplar.cs
… src/OpenTelemetry/Metrics/Exemplar/ExemplarReservoir.cs
… src/OpenTelemetry/Metrics/Exemplar/FixedSizeExemplarReservoir.cs
… src/OpenTelemetry/Metrics/Exemplar/ReadOnlyExemplarCollection.cs
… src/OpenTelemetry/Metrics/Exemplar/SimpleExemplarReservoir.cs
… src/OpenTelemetry/Metrics/Exemplar/SimpleFixedSizeExemplarReservoir.cs
… src/OpenTelemetry/Metrics/Metric.cs
… src/OpenTelemetry/Metrics/MetricPoint.cs
… src/OpenTelemetry/Metrics/MetricPointOptionalComponents.cs
… src/OpenTelemetry/Metrics/MetricReaderExt.cs
… src/OpenTelemetry/Metrics/MetricStreamConfiguration.cs
… src/OpenTelemetry/OpenTelemetry.csproj
… src/OpenTelemetry/ReadOnlyFilteredTagCollection.cs
… src/OpenTelemetry/ReadOnlyTagCollection.cs
… src/Shared/ThreadSafeRandom.cs
… test/OpenTelemetry.Tests/Metrics/AggregatorTestsBase.cs
… test/OpenTelemetry.Tests/Metrics/MetricExemplarTests.cs
… test/OpenTelemetry.Tests/Metrics/MetricTestsBase.cs