## Changes

### ExemplarReservoir

* `Offer` method signature no longer includes the "index" parameter which is not part of the spec. Instead it now accepts `struct ExemplarMeasurement` which only exposes the value & tags (to match spec) but also tracks explicit histogram index as an internal thing.
* `Collect` method signature no longer include "actualTags" or "reset" which are not part of the spec. The "reset" flag has become `bool ResetOnCollect` property on `ExemplarReservoir`. Tag filtering is now handled by `struct ReadOnlyFilteredTagCollection` which is exposed on `Exemplar`. Return type changed from `Exemplar[]` to `struct ReadOnlyExemplarCollection`.

### Exemplar

* Return type of `FilteredTags` property is now `struct ReadOnlyFilteredTagCollection` instead of `List<KeyValuePair<string, object?>>?`.
* Now exposes `long LongValue` in addition to `double DoubleValue`.

### SimpleFixedSizeExemplarReservoir

* Default size changed from `1` to `Environment.ProcessorCount` which is allowed by the spec.

### Concurrency

* `ExemplarReservoir` methods made safe to be called concurrently (required by the spec). Also fixed a bunch of race conditions with collect. Need to prove this out more probably using Coyote tests (follow-up).
* `MetricPoint` had dual-implementations of all aggregation types which acquired locks when exemplars were in play. This exploded the size/complexity of the code while also destroying throughput. Now that `ExemplarReservoir` can be called concurrently all of that has been removed.

### Misc improvements

* `ReadOnlyExemplarCollection` when enumerated only returns `Exemplar`s which have been updated (`Timestamp != default`). Previously each exporter had to understand the behavior and check that.
* Removed nasty Linq & allocations from `Exemplar.FilteredTags`.
* Introduced internal `FixedSizeExemplarReservoir` to house the shared logic both `SimpleFixedSizeExemplarReservoir` & `AlignedHistogramBucketExemplarReservoir` share.

## Benchmarks

* When `Exemplar`s are NOT enabled, perf is indistinguishable from before:

Before:

| Method         | AggregationTemporality | Mean     | Error    | StdDev   | Allocated |
|--------------- |----------------------- |---------:|---------:|---------:|----------:|
| CounterHotPath | Cumulative             | 10.33 ns | 0.128 ns | 0.107 ns |         - |
| CounterHotPath | Delta                  | 10.85 ns | 0.169 ns | 0.158 ns |         - |

After:

| Method         | AggregationTemporality | Mean     | Error    | StdDev   | Allocated |
|--------------- |----------------------- |---------:|---------:|---------:|----------:|
| CounterHotPath | Cumulative             | 10.48 ns | 0.136 ns | 0.127 ns |         - |
| CounterHotPath | Delta                  | 10.71 ns | 0.196 ns | 0.184 ns |         - |

* The `Exemplar` benchmarks show slight improvement:

Before:

| Method                    | ExemplarFilter | Mean     | Error   | StdDev  | Allocated |
|-------------------------- |--------------- |---------:|--------:|--------:|----------:|
| HistogramNoTagReduction   | AlwaysOff      | 176.0 ns | 1.30 ns | 1.15 ns |         - |
| HistogramWithTagReduction | AlwaysOff      | 168.7 ns | 3.40 ns | 3.91 ns |         - |
| HistogramNoTagReduction   | AlwaysOn       | 214.2 ns | 4.31 ns | 6.84 ns |         - |
| HistogramWithTagReduction | AlwaysOn       | 230.4 ns | 3.77 ns | 3.52 ns |         - |
| HistogramNoTagReduction   | HighValueOnly  | 192.7 ns | 1.91 ns | 1.59 ns |         - |
| HistogramWithTagReduction | HighValueOnly  | 188.6 ns | 3.72 ns | 4.56 ns |         - |

After:

| Method                    | ExemplarFilter | Mean     | Error   | StdDev  | Allocated |
|-------------------------- |--------------- |---------:|--------:|--------:|----------:|
| HistogramNoTagReduction   | AlwaysOff      | 170.3 ns | 1.54 ns | 1.44 ns |         - |
| HistogramWithTagReduction | AlwaysOff      | 164.3 ns | 1.88 ns | 1.67 ns |         - |
| HistogramNoTagReduction   | AlwaysOn       | 221.7 ns | 3.29 ns | 3.08 ns |         - |
| HistogramWithTagReduction | AlwaysOn       | 214.4 ns | 3.11 ns | 2.91 ns |         - |
| HistogramNoTagReduction   | HighValueOnly  | 185.8 ns | 3.12 ns | 2.92 ns |         - |
| HistogramWithTagReduction | HighValueOnly  | 177.1 ns | 3.29 ns | 3.23 ns |         - |

* Stress test shows massive gains:

Those `Exemplar` benchmarks above are very misleading. The `Exemplar` implementation before this PR forces all aggregation types into locking mode inside `MetricPoint`. This _destroys_ throughput for all non-histograms. That really won't show up on benchmarks because that is one thread hammering measurements.

Here are some stress test results:

Before:

* Total Runaway Time (seconds) 300
* Total Loops: 4,080,616,288
* Average Loops/Second: 13,613,263
* Average CPU Cycles/Loop: 511
* GC Total Allocated Bytes: 15181888

After:

* Total Runaway Time (seconds) 300
* Total Loops: 9,105,818,639
* Average Loops/Second: 30,314,633
* Average CPU Cycles/Loop: 1,843
* GC Total Allocated Bytes: 15048856

That is running the stress test using a `Counter<long>` + a view filtering tags + `AlwaysOnExemplarFilter` + an OtlpExporter tweaked to send exemplars for counters transmitting every 1000ms.

This new `Exemplar` implementation blows away the old one due to the concurrency changes. The reduction in allocations is due to the `FilteredTags` improvements.

## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-dotnet/blob/main/CONTRIBUTING.md) guidelines followed (license requirements, nullable enabled, static analysis, etc.)
* [X] Unit tests added/updated
* [X] Appropriate `CHANGELOG.md` files updated for non-trivial changes
* [ ] Changes in public API reviewed (if applicable)